# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Mosaic API
  version: 1.0.0
  description: |
    Zero-knowledge encrypted photo gallery API.
    
    The server never sees plaintext photos or metadata—all encryption/decryption 
    happens client-side. The server stores encrypted blobs and facilitates 
    synchronization between clients.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api
    description: API base path (behind reverse proxy)
security:
  - RemoteUser: []
...
paths:
...
  /shards/{shardId}:
    get:
      operationId: downloadShard
      summary: Download encrypted shard
      tags: [Shards]
      parameters:
        - name: shardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Encrypted shard data
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /shards/{shardId}/meta:
    get:
      operationId: getShardMetadata
      summary: Get shard metadata
      description: Returns server-side metadata for an encrypted shard without exposing shard plaintext.
      tags: [Shards]
      parameters:
        - name: shardId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Shard metadata
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ShardMetadata' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /shards:
    post:
      operationId: createShardUpload
      summary: Initialize shard upload (Tus protocol)
      description: |
        Creates a new shard upload. Returns Tus endpoint for resumable upload.
        Shard remains in PENDING status until linked to a manifest.
      tags: [Shards]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShardRequest'
      responses:
        '201':
          description: Upload initialized
          headers:
            Location:
              description: Tus upload URL
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShardCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '413':
          description: Shard too large (max 6MB)

  # ===========================================================================
  # Uploads (Tus 1.0.0)
  # ===========================================================================
...
  /s/{linkIdBase64}/shards/{shardId}:
    get:
      operationId: getShareLinkShard
      summary: Download encrypted shard via share link (anonymous)
      description: |
        Downloads an encrypted shard blob via share link without authentication.
        The shard must belong to a manifest in the linked album.
      tags: [Anonymous Access]
      security: []
      parameters:
        - name: linkIdBase64
          in: path
          required: true
          schema:
            type: string
          description: URL-safe base64 encoded link ID
        - name: shardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Shard UUID
      responses:
        '200':
          description: Encrypted shard blob
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          description: Link expired or revoked

  # ===========================================================================
  # Admin - Settings
  # ===========================================================================
...
components:
...
  securitySchemes:
    RemoteUser:
      type: apiKey
      in: header
      name: Remote-User
      description: |
        User identity from trusted reverse proxy.
        Only accepted from configured trusted IP ranges.
    cookieAuth:
      type: apiKey
      in: cookie
      name: mosaic_session
      description: |
        LocalAuth session cookie. The backend sets this cookie as HttpOnly with Path=/api,
        SameSite=Lax for development/testing HTTP, and Secure + SameSite=Strict in production.

  # ===========================================================================
  # Parameters
  # ===========================================================================
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Unauthorized:
      description: Authentication required
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Forbidden:
      description: Insufficient permissions
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
  schemas:
...
    ProblemDetails:
      type: object
      description: RFC 7807 problem details error response.
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        correlationId:
          type: string
          description: Request correlation identifier when available.
      additionalProperties: true
...
    ShardMetadata:
      type: object
      properties:
        id: { type: string, format: uuid }
        sizeBytes: { type: integer, format: int64 }
        status: { type: string }
        statusUpdatedAt: { type: string, format: date-time }
        sha256: { type: string }
...
    CreateShardRequest:
      type: object
      required:
        - sizeBytes
      properties:
        sizeBytes:
          type: integer
          format: int64
          minimum: 1
          maximum: 6291456
          description: Expected shard size in bytes (max 6MB)
    ShardCreated:
      type: object
      required:
        - id
        - uploadUrl
      properties:
        id:
          type: string
          format: uuid
        uploadUrl:
          type: string
          description: Tus upload endpoint URL

    # -------------------------------------------------------------------------
    # Share Links
    # -------------------------------------------------------------------------
...